python/requirements.txt
azure-storage
